Fix a slice ordering bug.
When the timeline slice group was constructed slices with the same start time
were being considered in the wrong order.  This could cause them to be drawn
with the wrong nesting order or on top of each other if they also had the same
duration.

BUG=62

Review URL: https://codereview.appspot.com/6575057

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@172 0e6d7f2b-9903-5b78-7403-59d27f066143