Add support for parsing kernel events with TGIDs
This change adds support for parsing kernel events that include a TGID column.
Because this new event line format changes the index of the regex groups, the
different event formats are now handled using a parsing function rather than a
regex.  The new parsing functions return an object with named fields so that
the order does not matter.  The various places that used the regex group array
have been updated to use the named fields instead.

Review URL: https://codereview.appspot.com/7388047

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@217 0e6d7f2b-9903-5b78-7403-59d27f066143