Don't rebuild deps and about_tracing on every request.
Currently, both deps and about_tracing will get regenerated for
every JS test file included in tests.html. This causes a full reload
of the tests page to take ~9-10 seconds.

With this patch, we'll rebuild the deps and about_tracing files at most
once every 5 seconds. This is fast enough that all of the JS files in
tests.html get loaded and try to load their deps requirement. This
takes the loading down to ~4-5 seconds for a full reload.

Review URL: https://codereview.appspot.com/8187043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@245 0e6d7f2b-9903-5b78-7403-59d27f066143