Fix category default selection issue.
If the category was selected by default it would not be included
in the list of categories sent to chrome because we'd only look
for categories in settings. This updates the code to look at the
actual selected inputs and sends those through to chrome.

BUG=https://code.google.com/p/trace-viewer/issues/detail?id=146

Review URL: https://codereview.appspot.com/8372043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@251 0e6d7f2b-9903-5b78-7403-59d27f066143