Add support for kernel function traces.
This change adds support for parsing the kernel function entry/return events
'graph_ent' and 'graph_ret'.  It also includes a way of dealing with the fact
that because userland tracing is done by writing to a /sys/ file, the kernel
and userland call tracing may not nest properly.

Review URL: https://codereview.appspot.com/7888046

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@252 0e6d7f2b-9903-5b78-7403-59d27f066143