Add history support to the analysis view.
Currently if the back button is used we'll lose your trace data
as chrome navigates back to before chrome://tracing was accessed.

With this CL, we will push the current selection onto an internal
array which will be used to restore the previous selection if
you use the back button. If you make a new selection, it is also
possible to navigate back to the previous selection.

Selections are not highlighted in the timeline view as you navigate
through the history.

This will use more memory as we will hold onto each selection as
you navigate.

BUG=
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/9111043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@312 0e6d7f2b-9903-5b78-7403-59d27f066143