Fix loading of non-JSON data files.
When loading trace files we need to make sure that the data is
actually JSON before attempting to parse. If the data isn't
JSON, e.g. Android systrace or v8 log files, we provide the raw
string as the trace events and the importers will do the right
thing.

BUG=150
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/9118044

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@313 0e6d7f2b-9903-5b78-7403-59d27f066143