Enable toggling the grid off.
If you press the 'g' or 'G' button we will enable the marker grid.
Currently there is no way to remove the grid, which can be annoying
when you just want to check something and end with all these red
lines on the screen.

This CL will disable the grid markers if you press the same key
again while on the same starting selection.

BUG=
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/9268049

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@396 0e6d7f2b-9903-5b78-7403-59d27f066143