Display name of loaded file in trace viewer.
When a file is loaded, if the name is provided, we will display the
name in the trace viewer header.

Currently, this will only display names of dropped files. When a file is
loaded using the 'Load' button the name will not be displayed until the
Chromium side of this patch is committed.

BUG=108
R=johnjbarton@chromium.org, nduca@chromium.org

Review URL: https://codereview.appspot.com/9329043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@413 0e6d7f2b-9903-5b78-7403-59d27f066143