Mutate objects immediately while creating implicit snapshots.
We used to discover snapshots, memoize them, and then later patch up
the objects that contained them. This was a mess because we'd get confused
and patch up an the object's original args instead of the object.

This patch creates object snapshots as we encounter them, replacing
them inline, and then continuing the search for more implicit snapshots in
the newly-created snapshot.

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@429 0e6d7f2b-9903-5b78-7403-59d27f066143