Auto delete object instances at end of recording
Traces typically contain sequences of objects that are crated, snapshotted,
but never deleted, e.g.:
  id created 0x1000
  snapshot 0x1000
  snapshot 0x1000
  snapshot 0x1000

When this happens, we consider this object undeleted, and it has a deletionTs
of MAX_VALUE. This adds a pass after finalizing import that marks these with
a deletionTs of the maxValue of the model. This improves trace readability.

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@532 0e6d7f2b-9903-5b78-7403-59d27f066143