Allow tracks to be hidden by viewers
Some object views are registered, but not with the intention of giving them
a track in the UI. Now, when a view is registered, you can pass it an options
object, on which you can pass showInTrackView=false to prevent it from
appearing in the track view.

We use this to make the tile track not show up.

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@534 0e6d7f2b-9903-5b78-7403-59d27f066143