Cleanup importer requires.
Currently the v8.log importer doesn't work in the trace-viewer example
file because it wasn't included in profiling_view. This moves the importers
over to a tracing/importer JS file that does the requiring (similar to how
cc.js does the cc requires).

This also removes the requiring of the importers in timeline_view as,
I believe, this should only happen in profiling_view.

BUG=
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/9698045

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@545 0e6d7f2b-9903-5b78-7403-59d27f066143