Remove settings caching of Category Filter options.
When using the category filter, it causes a lot of confusion to
have the categories disabled over test runs. Being able to hide them
as looking at the results has been useful, but it causes a lot of
confusion when the tracks don't apepar on a new trace.

BUG=

Review URL: https://codereview.appspot.com/9974044

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@580 0e6d7f2b-9903-5b78-7403-59d27f066143