Fixes to base loader following unittest rework
Trace-viewer was broken. This was due to the trace_viewer.html's inline
script executing *before* the deferred scripts have come back. To fix this,
the trace_viewer inline script is pulled out to a js file, and appended
to the document as a defer.

This isn't quite enough. base.require for some reason doesn't actually
append scripts when they're required, but rather appends them in a strange
setTimeout that happens after the require is called. I really dont see the point of this... maybe it was once needed, but now we just need to append the scripts, and let the browser execute them in order.

With base.require directly appending the scripts, then all we have to do is
require timeline, then append a defer script with trace_viewer.js and things
just work.

TBR=dsinclair
R=johnjbarton

Review URL: https://codereview.appspot.com/10226043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@610 0e6d7f2b-9903-5b78-7403-59d27f066143