Make deps parsing failures visible in the browser
When deps parsing breaks, for exmaple due to a missing module or css file,
you get a blank test and you have to go back to the console where your
devserv er is running to figure out what is wrong. This puts a big red error
up in the test file with the failure.

https://dl.dropboxusercontent.com/u/58819336/DepsErr.png

R=dsinclair@chromium.org

Review URL: https://codereview.appspot.com/9766045

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@619 0e6d7f2b-9903-5b78-7403-59d27f066143