Allow decorate to take parameters.
This allows us to pass argument to the constructors of UI elements and those
args will be proxied down to the decorate() method when the object is
constructed.

If you inherit from a base class, you will need to list all of the parent
classes decorate arguments before yours. All args are passed to all decorate
methods in the chain.

BUG=
TEST=Unit tests pass; manually inspected all instantiate tests; tested lthi_cats.json in trace-viewer;
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/10638043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@663 0e6d7f2b-9903-5b78-7403-59d27f066143