Add Skia ops list view to the picture debugger
This is the first iteration of the new SkiaOps view in the picture
debugger. This patch adds a simple list view of the Skia ops
in place of the previous generic picture data view.

A screenshot of the new view can be seen at: pr.gg/ops.png

BUG=
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/10704044

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@664 0e6d7f2b-9903-5b78-7403-59d27f066143