Move the "deviceViewport" from the quad_view_viewport to the quad_stack.
The deviceViewport is part of the data we show in the quad_viewer, it is
a reference 'layer', not part of the camera apparatus. By moving it to
the quad_stack we avoid confusion and avoid putting the values onto the
quad_view_viewport for the single purpose of retriving them.

BUG=None
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/10751044

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@725 0e6d7f2b-9903-5b78-7403-59d27f066143