DrawingContainer controls the draw of child tracks.
With this change, every Track has a draw method. The reason for this is, once
we start implementing instant events we'll need to be able to draw to the
canvas from Thread and TraceModel tracks due to the three instant scopes.

This removes the callback and rAF code from DrawableTrack (which should probably
be renamed but I can't think of a good name at the moment). The rAF is added
in the DrawingContainer and it triggers a draw call on each of it's children.

This will be extended in a future CL to have a drawInstantEvents method that is
also called.

BUG=trace-viewer:297
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/11319043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@742 0e6d7f2b-9903-5b78-7403-59d27f066143