Introduce "Chrome" viewport.
This patch switches the viewport view from a grey mask to Chrome's
chrome. This allows us to reduce canvas sizes and remove raster
scaling in a followup patch.

Most of the changes are in the new rect_view.js which decorates
a Rect. QuadStack now creates a RectView to style the world rect (called worldViewportRectView).

Note: the images used have been made as small as possible:
1) The sizes have been reduced by 1/2 from their native resolution.
2) The images have been hand-trimmed to be as small as possible.

A preview of this change is available at http://pr.gg/tracechrome.png

BUG=
R=johnjbarton@chromium.org, nduca@chromium.org

Review URL: https://codereview.appspot.com/11243043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@756 0e6d7f2b-9903-5b78-7403-59d27f066143