Update presubmit to detect missing URL entries in GYP.
We need to make sure that all of the CSS url() entries exist in src/images/
and are listed in the GYP file. Otherwise, Chromium won't build. This adds
and extra presubmit check that gathers all url() entries and everything
in src/images and looks for differences.

BUG=
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/11521044

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@767 0e6d7f2b-9903-5b78-7403-59d27f066143