Introduce PictureAsCanvas and remove PictureAsImage
Previously, we stored raster results as images which meant they had to
be encoded as pngs, only to be immediately decoded. This was slow! This
patch switches to using canvas for the picture raster result, improving
the speed of the frame viewer (toDataUrl was expensive).

This patch removes quad.backgroundRasterData which appears to not
be used anymore.

In a followup patch the picture debugger will be changed to use
a canvas instead of a background image.

BUG=
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/11563043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@769 0e6d7f2b-9903-5b78-7403-59d27f066143