Automatically add tests for low/hi dpi.
This is an alternate approach to the hi/low DPI testing. Instead of having
a button that we have to remember to click, we add an anotation to
tests that need to be DPI aware. {dpiAware: true}. This will tell the
test framework to generate two tests, one at low DPI and one at hi DPI. This
way, we will always be running the hi/low DPI tests.

BUG=
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/12282043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@821 0e6d7f2b-9903-5b78-7403-59d27f066143