Add importer code for flow events.
This adds the code to the importer to correctly load up flow events. The events
are added into their thread as though they were a regular slice. They are also
added into the trace_model so we'll be able to iterate and add all of the flow
arrows to the UI.

BUG=trace-viewer:121
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/12350043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@822 0e6d7f2b-9903-5b78-7403-59d27f066143