Draw arrows on flow lines.
Currently for flow events we assume the flow of events is left to right, the
same as the flow of time. This assumption is not necessarily obvious to people
who are new to trace viewer.

With this CL we will draw arrows on the end of the flow lines if there is
enough space so it doesn't look too wonky.

BUG=trace-viewer:367
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/12558043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@835 0e6d7f2b-9903-5b78-7403-59d27f066143