Make alternate mouse modes work regardless of the current mode.
This change adds a modifier system to MouseModeSelector that allows for
setting modifiers on modes. When the modifier is used the
MouseModeSelector will temporarely switch to this mode, regardless of
the current mode.

modifiers:
hold SHIFT for SELECTION
hold SPACE for PAN
hold CTRL/CMD for ZOOM

BUG=406
R=dsinclair@chromium.org

Review URL: https://codereview.appspot.com/13130044

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@913 0e6d7f2b-9903-5b78-7403-59d27f066143