Use min when merging op timings
A lot of noise in timings comes from CPU scheduling noise, per results from
@ernstm when working on rasterize_and_record_benchmark. I think @benchen also
derived this result in his early bench_pictures work. This patch takes the
minimum of the op costs instead of averaging.

I went into the new skia_debugger.html and reloaded the page a bunch and
indeed, the noise is reduced. I'd like to increase the iteration count to 8 as
that seemed to help a lot, too, but I think its better to that as another
effort, as the debugger commands are already pretty slow.

R=pdr@google.com
CC=ernstm,fmalita

Review URL: https://codereview.appspot.com/13420044

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@947 0e6d7f2b-9903-5b78-7403-59d27f066143