Add SelectionState to tracing Event
This change replaces the selected property on the tracing Event with the
enum SelectionState:

  var SelectionState = {
    NONE: 0,
    SELECTED: 1,
    HIGHLIGHTED: 2,
    DIMMED: 3
  };

TimelineTrackView gets another Selection highlight, which is used to store
highlighted events. Getters, setters and other methods for setting/clearing
selection and highlight are added to TimelineTrackView.

TraceModel.iterateAllEvents() is used to change the state on all events in
the model.

BUG=74
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/13493043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@949 0e6d7f2b-9903-5b78-7403-59d27f066143