Reset flex settings for the picture_view.
If we don't change this setting then the .analysis-view > * setting will take
effect which causes the top chart to not have a scrollbar, instead the right
pane becomes the width of the top graph. This can be very large for large
skps.

With this change, the top graph correctly gets a scrollbar and our drawing of
the skp appears onscreen instead of way off to the right where we have to scroll
to see it.

BUG=
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/13517046

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@1017 0e6d7f2b-9903-5b78-7403-59d27f066143