Make 'm' key put markers around the current selection
Its often the case that we want to put a mark around the current selection in order to then study that time frame on other threads. You can accomplish this using the marker tool, but when you're power-using, thats a whole lotta clicks. This shortcuts it. )

BUG=484
R=dsinclair@chromium.org

Review URL: https://codereview.appspot.com/14990043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@1038 0e6d7f2b-9903-5b78-7403-59d27f066143